Revamp use of assessment label
