+++ apps/prairielearn/src/models/assessment.ts
@@ -1,8 +1,9 @@
 import { z } from 'zod';
 
 import { loadSqlEquiv, queryOptionalRow, queryRow } from '@prairielearn/postgres';
+import * as sqldb from '@prairielearn/postgres';
 
-import { type Assessment, AssessmentSchema } from '../lib/db-types.js';
+import { type Assessment, AssessmentSchema, IdSchema } from '../lib/db-types.js';
 
 const sql = loadSqlEquiv(import.meta.url);
 
@@ -19,4 +20,18 @@ export async function selectOptionalAssessmentById(
   assessment_id: string,
 ): Promise<Assessment | null> {
   return await queryOptionalRow(sql.select_assessment_by_id, { assessment_id }, AssessmentSchema);
+}
+
+const AssessmentInfoSchema = z.object({
+  assessment_label: z.string(),
+  course_instance_id: IdSchema,
+  course_id: IdSchema,
+});
+
+export async function selectAssessmentInfoForJob(assessment_id: string) {
+  return await sqldb.queryRow(
+    sql.select_assessment_info_for_job,
+    { assessment_id },
+    AssessmentInfoSchema,
+  );
 }
